writer profile now made up of separate components
