+++ src/components/writer-profile/writerProfile.js
@@ -1,98 +1,32 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { StylesProvider, withStyles } from "@material-ui/core/styles";
-import Paper from "@material-ui/core/Paper";
 import { useStyles } from "./writerProfile.styles.js";
-import { useStyles as workStyles } from "../WriterOnboardingForms/WriterForm.styles";
-import AccountCircleIcon from "@material-ui/icons/AccountCircle";
-import PropTypes from "prop-types";
 import Loader from "../loader/Loader.js";
-import {
-  EditButton,
-  EditProfile,
-} from "../EditProfileForms/EditProfileForms.js";
-import {
-  Typography,
-  Button,
-  Tabs,
-  Tab,
-  Link,
-  Box,
-  Card,
-  CardContent,
-} from "@material-ui/core";
+import { EditProfile } from "../EditProfileForms/EditProfileForms.js";
+import { Grid } from "@material-ui/core";
 import {
   getWriterInfo,
   updateWriterProfile,
 } from "../../store/actions/profileActions.js";
+import WriterEducation from "./WriterEducation.js";
+import WriterServices from "./WriterServices.js";
+import WorkHistory from "./WorkHistory.js";
+
+//shared imports currently located in applicant-profile folder
 import BioCard from "../applicant-profile/BioCard.js";
-
-const GlobalCSS = withStyles({
-  "@global": {
-    ".MuiPaper-elevation2": {
-      width: "80%",
-      height: "auto",
-      marginTop: "5%",
-      marginBottom: "5%",
-      marginLeft: "10%",
-    },
-  },
-})(() => null);
-
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`vertical-tabpanel-${index}`}
-      aria-labelledby={`vertical-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box p={3}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired,
-};
-
-function a11yProps(index) {
-  return {
-    id: `vertical-tab-${index}`,
-    "aria-controls": `vertical-tabpanel-${index}`,
-  };
-}
+import LeftPanel from "../applicant-profile/LeftPanel.js";
 
 const WriterProfile = (props) => {
   const dispatch = useDispatch();
   const classes = useStyles();
-  const workClasses = workStyles();
 
   //Redux
   const writer = useSelector((state) => state.profileInfo.profileDetails);
   const userId = useSelector((state) => state.login.userId);
-  useEffect(() => {
-    dispatch(getWriterInfo(userId));
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [userId]);
   const userType = useSelector((state) => state.login.usertype);
   const isEditing = useSelector((state) => state.profileInfo.isEditing);
   const isLoading = useSelector((state) => state.profileInfo.isLoading);
-  const profileId = useSelector(
-    (state) => state.profileInfo.profileDetails.writer_id
-  );
-  const viewerId = useSelector((state) => state.login.userId);
 
-  const [value, setValue] = React.useState(0);
   const [profile, setProfile] = useState({
     first_name: writer.first_name,
     last_name: writer.last_name,
@@ -110,11 +44,6 @@ const WriterProfile = (props) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [userId]);
   //
-
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
-
   const editHandleChange = (event) => {
     setProfile({
       ...profile,
@@ -129,54 +58,18 @@ const WriterProfile = (props) => {
   };
 
   return (
-    <StylesProvider>
-      <GlobalCSS />
+    <div className={classes.root}>
       {writer && !isLoading ? (
-        <Paper classes={{ root: classes.rootPaper }} elevation={3}>
-          {" "}
-          <div>
-            <AccountCircleIcon
-              classes={{ root: classes.rootIcon }}
-              fontSize="large"
-            />
-            {writer && (
-              <div className={classes.userName}>
-                {" "}
-                {writer.first_name} {writer.last_name}
-              </div>
-            )}
-            {writer && (
-              <>
-              <div className={classes.leftPanel}>
-                <Link
-                  classes={{ root: classes.rootLink }}
-                  href={`http://${writer.website}`}
-                >
-                  {writer.website}
-                </Link>
-
-                <Button
-                  classes={{ root: classes.rootButton, label: classes.labelButton }}
-                  variant="contained"
-                  color="primary"
-                  href="#contained-buttons"
-                >
-                  Direct Message
-                </Button>
-                <div>  
-                <EditButton 
-                  classes={classes.editButton}
-                  viewerId={viewerId} 
-                  profileId={profileId} 
-                />
-                {/*This is only rendered if the viewerId matches the ProfileId...only the profile owner has the option to edit their profile.*/}
-                </div>
-              </div>
-              </>
-            )}
-          </div>
-          {writer && (
-            <>
+        <div>          
+          <Grid className={classes.profile}>
+            
+            <div className={classes.leftpanel}>
+              <LeftPanel 
+                profileDetails={writer}
+              />
+            </div>
+            
+            <div>
               {isEditing === true ? (
                 <EditProfile
                   profile={profile}
@@ -185,98 +78,31 @@ const WriterProfile = (props) => {
                   userType={userType}
                 />
               ) : (
-                <div>
                   <BioCard 
                     profileDetails={writer}
                   />
-                </div>
               )}
-            </>
-          )}
-          <div></div>
-          <h3 className={classes.userEducation}>
-            Background:
-            <div className={classes.educationText}>
-              Education:<div className={classes.bodyText}>USC 2010-2014</div>
             </div>
-          </h3>
-          <div></div>
-          <div className={classes.userServices}>
-            <h3 className={classes.userEducation}>Services Offered:</h3>
-            <Paper>
-              <Tabs
-                orientation="vertical"
-                variant="scrollable"
-                value={value}
-                onChange={handleChange}
-                aria-label="Vertical tabs example"
-                className={classes.tabs}
-              >
-                <Tab label="Grant Writing" {...a11yProps(0)} />
-                <Tab label="Grant Research" {...a11yProps(1)} />
-              </Tabs>
-              <TabPanel value={value} index={0}>
-                Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quos
-                blanditiis tenetur unde suscipit, quam beatae rerum inventore
-                consectetur, neque doloribus, cupiditate numquam dignissimos
-                laborum fugiat deleniti? Eum quasi quidem quibusdam.
-              </TabPanel>
-              <TabPanel value={value} index={1}>
-                This is just here to show you this works.
-              </TabPanel>
-            </Paper>
-          </div>
-          <h3 className={classes.finalGrid}>
-            Work History:
-            {writer &&
-              writer.workHistory &&
-              writer.workHistory.map((writersWorkHistory) => (
-                <Card
-                  className={workClasses.cardRoot}
-                  key={writersWorkHistory.id}
-                  variant="outlined"
-                >
-                  <CardContent>
-                    <Typography
-                      className={workClasses.title}
-                      color="textSecondary"
-                      gutterBottom
-                      data-testid="company-header"
-                    >
-                      Company
-                    </Typography>
-                    <Typography variant="h5" component="h2">
-                      {writersWorkHistory.company}
-                    </Typography>
-                    <Typography
-                      className={workClasses.pos}
-                      color="textSecondary"
-                      data-testid="position-header"
-                    >
-                      Position: {writersWorkHistory.position}
-                    </Typography>
-                    <Typography variant="body2" component="p">
-                      Start Date: {writersWorkHistory.start_date}
-                      <br />
-                      {writersWorkHistory.current_position === "true"
-                        ? `Current Position`
-                        : `End Date: ${writersWorkHistory.end_date}`}
-                    </Typography>
-                    <Typography
-                      className={workClasses.pos}
-                      color="textSecondary"
-                    >
-                      responsibilities: {writersWorkHistory.responsibilities}
-                    </Typography>
-                  </CardContent>
-                </Card>
-              ))}
-          </h3>
-        </Paper>
+            
+            <div>
+            <WriterEducation />
+            </div>
+          
+            <div>  
+            <WriterServices />
+            </div>
+            <div>            
+            <WorkHistory 
+              writer={writer}
+            />
+            </div>
+            
+          </Grid> 
+        </div>
       ) : (
         <Loader />
       )}
-    </StylesProvider>
+    </div>
   );
 };
